Minor tests with labels.dtype in code/train_val_test_utilities.py.

+++ code/train_val_test_utilities.py
@@ -97,7 +97,9 @@ def run_model(model, dataloader, mode, device, optimizer=None, class_specific=Tr
                 print(labels_.shape)
                 print(labels_.dtype)
                 print(labels_)
-                prototypes_of_correct_class = torch.t(model.prototype_class_identity[:,labels]).to(device)
+                
+                # Note: We had to convert labels to torch.int64 to allow slicing
+                prototypes_of_correct_class = torch.t(model.prototype_class_identity[:,labels.to(torch.int64)]).to(device)
                 exit()
                 inverted_distances, _ = torch.max((max_dist - min_distances) * prototypes_of_correct_class, dim=1)
                 cluster_cost = torch.mean(max_dist - inverted_distances)
